superlingo_be/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
requests
google-generativeai
google-cloud-texttospeech # Added Google Cloud TTS library
google-cloud-speech